LLM_KopAI/requirements.txt
google-generativeai
python-dotenv
google-cloud-secret-manager
Pillow